Restrict population to currently registered or deceased
Include people who are registered at the end of the study period, so
we aren't looking at codes for people moving practices, which may skew
the results. However, this excludes deceased people, so we also include
anyone who has died during the study period. This ensures we capture any
end-of-life related codes.